gsAssembler

The methods in gsAssembler contain methods used for assembly.

Function Documentation

Gismo.GaussRule — Function

Returns a Gauss rule for numerical integration

Arguments

numNodes::Int: the number of points in each direction

digits::Int: the number of digits of precision for the rule (optionals)

Gismo.GaussRule — Function

Returns a Gauss rule for numerical integration

Arguments

d::Int: the dimension of the rule

numNodes::Array{Int}: a vector of length d with the number of points in each direction

digits::Int: the number of digits of precision for the rule (optionals)

Gismo.LobattoRule — Function

Returns a Lobatto rule for numerical integration

Arguments

numNodes::Int: the number of points in each direction

digits::Int: the number of digits of precision for the rule (optionals)

Gismo.LobattoRule — Function

Returns a Lobatto rule for numerical integration

Arguments

d::Int: the dimension of the rule

numNodes::Array{Int}: a vector of length d with the number of points in each direction

digits::Int: the number of digits of precision for the rule (optionals)

Gismo.dim — Method

Returns the dimension of the quad rule

Arguments

qr::QuadRule: the quad rule

Gismo.mapTo — Method

Maps the quad rule to a given interval

Arguments

qr::QuadRule: the quad rule

startVal::Cdouble: the lower bound of the interval

endVal::Cdouble: the upper bound of the interval

Gismo.mapTo — Method

Maps the quad rule to a given interval

Arguments

qr::QuadRule: the quad rule

lower::Vector{Cdouble}: the lower bound of the interval

upper::Vector{Cdouble}: the upper bound of the interval

Gismo.numNodes — Method

Returns the number of nodes

Arguments

qr::QuadRule: the quad rule

Gismo.QuadRule — Type

A struct that represents a quadrature rule