gsCore

Function Documentation

Gismo.actives — Method

Returns the actives of a basis ...

Arguments

obj::Basis: a Gismo Basis

u::Matrix{Cdouble}: a matrix of points

...

Gismo.addPatch — Method

Returns the basis of a MultiPatch ...

Arguments

obj::MultiPatch: a Gismo MultiPatch

...

Gismo.basis — Method

Returns the basis of a geometry ...

Arguments

obj::Geometry: a Gismo Geometry

...

Gismo.basis — Method

Returns the basus of a MultiBasis ...

Arguments

obj::MultiBasis: a Gismo MultiBasis

i::Int: the index of the basis

...

Gismo.basis — Method

Returns the coefficients of a MultiPatch ...

Arguments

obj::MultiPatch: a Gismo MultiPatch

...

Gismo.closest — Function

Returns the closest point of a geometry ...

Arguments

obj::Geometry: a Gismo Geometry

x::Matrix{Cdouble}: a matrix of points

accuracy::Cdouble=1e-6: the accuracy of the computation

...

Gismo.coefs — Method

Returns the coefficients of a geometry ...

Arguments

obj::Geometry: a Gismo Geometry

...

Gismo.component — Method

Returns the component of a basis ...

Arguments

obj::Basis: a Gismo Basis

i::Cint: the index of the component

...

Gismo.degree — Method

Returns the degree of a basis ...

Arguments

obj::Basis: a Gismo Basis

i::Cint: the index of the component

...

Gismo.domainDim — Method

Returns the domain dimension of a basis ...

Arguments

object::Basis: a Gismo Basis

...

Gismo.domainDim — Method

Return the domain dimension of a geometry ...

Arguments

object::Geometry: a Gismo Geometry

...

Gismo.domainDim — Method

Adds a patch to a MultiPatch ...

Arguments

obj::MultiPatch: a Gismo MultiPatch

geom::Geometry: a Gismo Geometry

...

Gismo.eval — Method

Returns the evaluation of a basis ...

Arguments

obj::Basis: a Gismo Basis

u::Matrix{Cdouble}: a matrix of points

...

Gismo.eval — Method

Returns the evaluation of a geometry ...

Arguments

obj::Geometry: a Gismo Geometry

u::Matrix{Cdouble}: a matrix of points

...

Gismo.invertPoints — Function

Inverts a set of points ...

Arguments

obj::Geometry: a Gismo Geometry

x::Matrix{Cdouble}: a matrix of points

accuracy::Cdouble=1e-6: the accuracy of the computation

...

Gismo.normal — Method

Returns the normal of a geometry ...

Arguments

obj::Geometry: a Gismo Geometry

u::Matrix{Cdouble}: a matrix of points

...

Gismo.numElements — Method

Returns the number of elements of a basis ...

Arguments

obj::Basis: a Gismo Basis

...

Gismo.patch — Method

Returns the coefficients of a MultiPatch ...

Arguments

obj::MultiPatch: a Gismo MultiPatch

...

Gismo.refine — Function

Refines a basis ...

Arguments

obj::Basis: a Gismo Basis

boxes::Matrix{Cdouble}: the boxes to refine (first column is the lower bound, second column is the upper bound)

...

Gismo.refineElements — Method

Refines a basis ...

Arguments

obj::Basis: a Gismo Basis

boxes::Vector{Cint}: the boxes to refine (in index format)

...

Gismo.size — Method

Returns the size of a basis ...

Arguments

obj::Basis: a Gismo Basis

...

Gismo.targetDim — Method

Returns the target dimension of a basis ...

Arguments

object::Basis: a Gismo Basis

...

Gismo.targetDim — Method

Returns the target dimension of a geometry ...

Arguments

object::Geometry: a Gismo Geometry

...

Gismo.targetDim — Method

Returns the target dimension of a MultiPatch ...

Arguments

object::MultiPatch: a Gismo MultiPatch

...

Gismo.uniformRefine — Function

Refines a basis ...

Arguments

obj::Basis: a Gismo Basis

numKnots::Cint=Int32(1): the number of knots to add

mul::Cint=Int32(1): the multiplicity of the knots

dir::Cint=Int32(-1): the direction of the refinement

...

Gismo.Basis — Type

Makes a Gismo Basis

Gismo.Geometry — Type

Makes a Gismo Geometry

Gismo.MultiBasis — Type

Makes a Gismo MultiBasis

Gismo.MultiPatch — Type

Makes a Gismo MultiPatch

gsCore

Function Documentation

gsNurbs

Function Documentation

gsHSplines

Function Documentation