gsMatrix

The methods in gsMatrix are an interface to G+Smo's linear algebra, supported by Eigen.

Function Documentation

Gismo.asMatrix — Method

Returns the matrix as a Julia matrix.

Arguments

object::EigenMatrix: the matrix

Gismo.asMatrixInt — Method

Returns the matrix as a Julia matrix.

Arguments

object::EigenMatrixInt: the matrix

Gismo.asVector — Method

Returns the matrix as a Julia vector.

Arguments

object::EigenMatrix: the matrix

Gismo.asVectorInt — Method

Returns the matrix as a Julia vector.

Arguments

object::EigenMatrixInt: the matrix

Gismo.cols — Method

Returns the number of columns in the matrix.

Arguments

object::EigenMatrixInt: the matrix

Examples

m = EigenMatrixInt(3,3)
print(Gismo.cols(m))
# output
3

Gismo.cols — Method

Returns the number of columns in the matrix.

Arguments

object::EigenMatrix: the matrix

Examples

m = EigenMatrix(3,3)
print(Gismo.cols(m))
# output
3

Gismo.data — Method

Returns a pointer to the data of the matrix.

Arguments

object::EigenMatrixInt: the matrix

Gismo.data — Method

Returns a pointer to the data of the matrix.

Arguments

object::EigenMatrix: the matrix

Gismo.rows — Method

Returns the number of rows in the matrix.

Arguments

object::EigenMatrixInt: the matrix

Examples

m = EigenMatrixInt(3,3)
print(Gismo.rows(m))
# output
3

Gismo.rows — Method

Returns the number of rows in the matrix.

Arguments

object::EigenMatrix: the matrix

Examples

m = EigenMatrix(3,3)
print(Gismo.rows(m))
# output
3

Gismo.setZero! — Method

Sets all the elements of the matrix to zero.

Arguments

object::EigenMatrixInt: the matrix

Gismo.setZero! — Method

Sets all the elements of the matrix to zero.

Arguments

object::EigenMatrix: the matrix

Gismo.toMatrix — Method

Returns the matrix as a Julia matrix (transfers ownership).

Arguments

object::EigenMatrix: the matrix

Gismo.toMatrixInt — Method

Returns the matrix as a Julia matrix (transfers ownership).

Arguments

object::EigenMatrixInt: the matrix

Gismo.toVector — Method

Returns the matrix as a Julia vector (transfers ownership).

Arguments

object::EigenMatrix: the matrix

Gismo.toVectorInt — Method

Returns the matrix as a Julia vector (transfers ownership).

Arguments

object::EigenMatrixInt: the matrix

Gismo.EigenMatrix — Type

EigenMatrix()

Creates an empty matrix.

Example

m = EigenMatrix()
# output