gsModeling

The methods in gsModeling contain methods for geometric modelling, such as fitting.

Function Documentation

Gismo.compute! — Function

compute!(fit::Fitting, lambda::Cdouble=0.0) Computes the least squares fit

Arguments

fit::Fitting: a fitting structure

lambda::Cdouble: the value to assign to the lambda ridge parameter

Gismo.computeErrors! — Method

computeErrors!(fit::Fitting) Computes the error for each point

Arguments

fit::Fitting: a fitting structure

Gismo.maxPointError — Method

maxPointError(fit::Fitting)::Cdouble Returns the maximum error obtained

Arguments

fit::Fitting: a fitting structure

Return

max_error::Cdouble: The maximum error obtained

Gismo.minPointError — Method

minPointError(fit::Fitting)::Cdouble Returns the smallest error obtained between all the points

Arguments

fit::Fitting: a fitting structure

Return

min_error::Cdouble: The minimum error obtained

Gismo.numPointsBelow — Method

numPointsBelow(fit::Fitting, threshold::Cdouble)::Int Returns the number of points where the error is below the threshold

Arguments

fit::Fitting: a fitting structure

threshold::Cdouble: The desired threshold

Return

number_pts_blw::Int: number of points where the error is below the given threshold

Gismo.parameterCorrection! — Function

parameterCorrection!(fit::Fitting, accuracy::Cdouble, maxIter::Int, tol0rth::Cdouble) Performs the parameters corrections step

Arguments

fit::Fitting: a fitting structure

accuracy::Cdouble: The desired accuracy

maxIter::Int: The desired number of iterations

tol0rth::Cdouble: The desired value of the tolleance

Gismo.pointWiseErrors — Method

pointWiseErrors(fit::Fitting)::Ptr{Cdouble} Returns the error obtained for each point

Arguments

fit::Fitting: a fitting structure

Return

errors::Ptr{Cdouble}: Pointer pointing to an array containing the error value for each point

Gismo.result — Method

result(fit::Fitting)::Geometry Returns a geometry from the fitting structure

Arguments

fit::Fitting: a fitting structure

Return

geom::Geometry: the desired geometry

Gismo.Fitting — Type

Creates a fitting structure