gsNurbs

Function Documentation

Gismo.BSpline — Method

BSpline(basis::Basis,coefs::Matrix{Cdouble})

Defines a B-spline geometry.

Arguments

basis::Basis: the basis

coefs::Matrix{Cdouble}: the coefficients

Gismo.BSplineBasis — Method

BSplineBasis(kv::KnotVector)

Defines a B-spline basis.

Arguments

kv::KnotVector: the knot vector

Gismo.Nurbs — Method

Nurbs(basis::Basis,coefs::Matrix{Cdouble})

Defines a NURBS geometry.

Arguments

basis::Basis: the basis

coefs::Matrix{Cdouble}: the coefficients

Gismo.NurbsBasis — Method

NurbsBasis(kv::KnotVector)

Defines a NURBS basis.

Arguments

kv::KnotVector: the knot vector

Gismo.TensorBSpline — Method

TensorBSpline(basis::Basis,coefs::Matrix{Cdouble})

Defines a tensor B-spline geometry.

Arguments

basis::Basis: the basis

coefs::Matrix{Cdouble}: the coefficients

Gismo.TensorBSplineBasis — Method

TensorBSplineBasis(kv1::KnotVector,kv2::KnotVector,kv3::KnotVector,kv4::KnotVector)

Defines a 4D tensor B-spline basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

kv3::KnotVector: the third knot vector

kv4::KnotVector: the fourth knot vector

Gismo.TensorBSplineBasis — Method

TensorBSplineBasis(kv1::KnotVector,kv2::KnotVector,kv3::KnotVector)

Defines a 3D tensor B-spline basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

kv3::KnotVector: the third knot vector

Gismo.TensorBSplineBasis — Method

TensorBSplineBasis(kv1::KnotVector,kv2::KnotVector)

Defines a 2D tensor B-spline basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

Gismo.TensorNurbs — Method

TensorNurbs(basis::Basis,coefs::Matrix{Cdouble})

Defines a tensor NURBS geometry.

Arguments

basis::Basis: the basis

coefs::Matrix{Cdouble}: the coefficients

Gismo.TensorNurbsBasis — Method

TensorNurbsBasis(kv1::KnotVector,kv2::KnotVector,kv3::KnotVector,kv4::KnotVector)

Defines a 4D tensor NURBS basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

kv3::KnotVector: the third knot vector

kv4::KnotVector: the fourth knot vector

Gismo.TensorNurbsBasis — Method

TensorNurbsBasis(kv1::KnotVector,kv2::KnotVector,kv3::KnotVector)

Defines a 3D tensor NURBS basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

kv3::KnotVector: the third knot vector

Gismo.TensorNurbsBasis — Method

TensorNurbsBasis(kv1::KnotVector,kv2::KnotVector)

Defines a 2D tensor NURBS basis.

Arguments

kv1::KnotVector: the first knot vector

kv2::KnotVector: the second knot vector

Gismo.size — Method

size(kv::KnotVector)

Returns the number of elements in the knot vector.

Arguments

kv::KnotVector: the knot vector

Examples

```jldoctest output=(false)

kv = KnotVector(Float64[0.,0.,0.,0.,0.5,1.,1.,1.,1.])

print(Gismo.size(kv))

# output

```

Gismo.KnotVector — Type

KnotVector(filename::String)

Arguments

filename::String: the name of the file containing the knot vector

KnotVector(a::Vector{Float64})

Arguments

a::Vector{Float64}: the knot vector

Examples

kv = KnotVector(Float64[0.,0.,0.,0.,0.5,1.,1.,1.,1.])
# output